Corrí las pruebas
esto dio
10
11
12
13
14
15
16
17
18
19
20
comparations taken, insert bubble merge quick
[ 523776 2096128 8386560 33550336 134209536 536854528 2147450880
8589869056 34359607296 137438691328 549755289600 ]
[ 1047552 4192256 16773120 67100672 268419072 1073709056 4294901760
17179738112 68719214592 274877382656 1099510579200 ]
[ 7558 16733 37766 82837 183647 401019 866501 1863925 3994105 8514011
18053762 ]
[ 56978 165227 207583 702288 1502090 2128544 7404287 17757498 28595019
56781319 116944048 ]
time taken, insert bubble merge quick
[ 13840 3611 9356 37202 156270 573278 2333256 9617377 37789367 150672981
602890661 ]
[ 15102 5926 16490 67893 279201 1138382 4704920 19285296 77243526
306599252 1358202665 ]
[ 2838 1324 2302 1247 2723 5305 11124 22697 44701 91068 235265 ]
[ 6737 510 661 1094 2255 3442 10670 25287 43259 83468 186039 ]

+++ src/t1/Useful.java
@@ -25,7 +25,7 @@ public class Useful {
 			System.out.println(A[i] + "\t" + B[i]);
 		}
 	}
-	public static void printArray(int[] A){
+	public static void printArray(long[] A){
 		System.out.print("[ ");
 		for (int i = 0; i < A.length; i++) {
 			System.out.print(A[i] + " ");
